refactor based on instanceStatus logic, add dialog
