+++ src/component/admin/billing/BillingGrid/BillingInfoButton/BillingInfoButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@mui/material';
-import { FC } from 'react';
+import { VFC } from 'react';
 
 const href = `mailto:elise@getunleash.ai?subject=Continue with Unleash&body=Hi Unleash,%0D%0A%0D%0A
 I would like to continue with Unleash.%0D%0A%0D%0A%0D%0A%0D%0A
@@ -16,17 +16,21 @@ Billing information:%0D%0A%0D%0A
 
 -- Thank you for signing up. We will upgrade your trial as quick as possible and we will grant you access to the application again. --`;
 
-export const BillingInfoButton: FC = ({ children }) => {
+interface IBillingButton {
+    update?: boolean;
+}
+
+export const BillingInfoButton: VFC<IBillingButton> = ({ update }) => {
     return (
         <Button
             href={href}
-            variant="contained"
+            variant={update ? 'outlined' : 'contained'}
             sx={{
                 width: '100%',
                 marginBottom: '12px',
             }}
         >
-            {children}
+            {update ? 'Update billing information' : 'Add billing information'}
         </Button>
     );
 };
